Merge all the work to main

+++ database/migrations/2024_06_06_115554_create_order_product_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignUlid('order_id')->constrained();
             $table->foreignIdFor(Product::class)->constrained();
             $table->foreignIdFor(ProductVariant::class)->nullable()->constrained();
-            $table->float('unit_price');
+            $table->integer('unit_price');
             $table->integer('quantity');
             $table->timestamps();
         });
